test(wasm): exercise terminal-width WASM tests in CI
The new terminal-width tests never ran in any CI job. The wasm.yml
"Run WASM unit tests" step only covered ./cmd/megaport, and the native
job skips js && wasm files by build tag.

Wiring them in first meant fixing a compile blocker: the
internal/base/output test binary couldn't build under js/wasm because
several native-only test files (stdout/stderr pipe capture) were
untagged, and one referenced a helper defined only in a !wasm file. Tag
those files !wasm and move the shared SimpleStruct/ComplexStruct
fixtures into an untagged file so both builds see them.

CI now runs the whole node-safe output wasm suite plus the
TestSetTerminalWidth tests in internal/wasm (the rest of that package
needs a browser host).
