+++ .github/workflows/wasm.yml
@@ -21,12 +21,15 @@ jobs:
       - name: Compile js/wasm target
         run: make wasm-build-guard
       - name: Run WASM unit tests
-        # Only ./cmd/megaport runs under the bare go_js_wasm_exec (node) runtime.
-        # The other WASM test packages need a browser host (fetch, JS bridge
-        # globals, stdin pipes) and are verified manually in-browser.
+        # Runs the WASM tests that are safe under the bare go_js_wasm_exec (node)
+        # runtime. ./cmd/megaport and the whole ./internal/base/output wasm suite
+        # run cleanly. In ./internal/wasm only the terminal-width tests are
+        # node-safe; the rest need a browser host (fetch, JS bridge globals,
+        # stdin pipes) and are verified manually in-browser.
         run: |
           export PATH="$PATH:$(go env GOROOT)/lib/wasm"
-          GOOS=js GOARCH=wasm go test -v -tags js,wasm ./cmd/megaport
+          GOOS=js GOARCH=wasm go test -v -tags js,wasm ./cmd/megaport ./internal/base/output
+          GOOS=js GOARCH=wasm go test -v -tags js,wasm -run 'TestSetTerminalWidth' ./internal/wasm
 
   smoke:
     name: Browser Fetch Smoke
