+++ internal/base/output/changes_test.go
@@ -1,3 +1,5 @@
+//go:build !wasm
+
 package output
 
 import (
